requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiogram==3.15.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.14.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
distlib==0.3.8
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnspython==2.7.0
fastapi==0.115.7
feedparser==6.0.11
ffmpeg-python==0.2.0
filelock==3.15.4
Flask==2.1.2
Flask-Cors==5.0.0
Flask-Migrate==3.1.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==2.5.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.12.0
future==1.0.0
gast==0.6.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-translate==3.18.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
gTTS==2.5.2
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.27.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
keras==3.7.0
kiwisolver==1.4.7
langdetect==1.0.9
libclang==18.1.1
magic-filter==1.0.12
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.1.0
mss==9.0.1
multidict==6.0.5
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
openai==1.54.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
plotly==5.24.1
proto-plus==1.25.0
protobuf==4.25.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pymongo==4.10.1
pynput==1.7.7
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.2.0
PyPDF2==3.0.1
pyTelegramBotAPI==4.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
python-telegram-bot==21.4
pytz==2024.2
PyYAML==6.0.2
pyzbar==0.1.9
redis==5.2.0
regex==2024.11.6
requests==2.31.0
rich==13.9.4
rsa==4.9
safetensors==0.5.2
schedule==1.2.2
seaborn==0.13.2
setuptools==75.8.0
sgmllib3k==1.0.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
speedtest-cli==2.1.3
spotipy==2.24.0
SQLAlchemy==1.4.23
sqlparse==0.5.2
starlette==0.45.3
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.5.0
tf_keras==2.16.0
tokenizers==0.21.0
torch==2.2.2
tqdm==4.67.0
transformers==4.48.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.1
uvicorn==0.34.0
virtualenv==20.26.3
Werkzeug==2.0.3
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
yarl==1.9.4
yt-dlp==2024.11.18
